+++ src/adlers/__init__.py
@@ -1,8 +1,9 @@
 from importlib.metadata import PackageNotFoundError, version
 
 from adlers.shared import ScaledDotProductAttention
+from adlers.time_series import ProbSparseAttention
 
-__all__ = ["ScaledDotProductAttention", "__version__"]
+__all__ = ["ProbSparseAttention", "ScaledDotProductAttention", "__version__"]
 
 try:
     __version__ = version("adlers")